+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Automatically retry fetching the spec if it doesn't work for any reason.
 * Fix array description display preference to be first--The Array description, and then only second--The array item schema description, so that the "more specific" description wins, even though the "items" description is more deeply nested. This aligns to the expected behavior of preference in the json schema.
 * Fix selecte drop down menu update trigger.
+* Add missing link to Tag display page.
+* Replace `replaceState` with `pushState` so forwards and backwards navigation actually work.
 
 ## 2.1
 * Add `x-locale` vendor extension to specify the locale of the spec.